Quick fix for mingw posix socket API

+++ examples/templates/posix_sockets.h
@@ -58,7 +58,7 @@ int open_nb_socket(const char* addr, const char* port) {
     if (sockfd != -1) fcntl(sockfd, F_SETFL, fcntl(sockfd, F_GETFL) | O_NONBLOCK);
 #else
     if (sockfd != INVALID_SOCKET) {
-        int iMode = 1;
+        u_long iMode = 1;
         ioctlsocket(sockfd, FIONBIO, &iMode);
     }
 #endif
